feat: 기존 UserInfoInputView를 UserInfoInputCellView로 변경 후 유저 정보를 입력받는 stackview와 버튼을 포함한 UserInfoInputView 생성
