+++ TheStreamers/TheStreamers/Scenes/SignInUp/UserInfoInputView.swift
@@ -11,23 +11,15 @@ class UserInfoInputView: UIView {
     
     private var type: InputType?
     
-    private lazy var textField: UITextField = {
-        let textField = UITextField()
-        textField.translatesAutoresizingMaskIntoConstraints = false
-        textField.placeholder = type?.placeholder()
-        textField.font = UIFont.systemFont(ofSize: 16)
-        textField.borderStyle = .none
-        return textField
+    private lazy var stackView: UserInfoInputStackView = {
+        UserInfoInputStackView(type: type ?? .signIn)
     }()
     
-    private lazy var button: UIButton? = {
-        guard type == InputType.emailForSignUp || type == InputType.code else {
-            return nil
-        }
+    private lazy var signInButton: MainButton = {
         let button = MainButton()
-        button.setTitle(type?.buttonTitle(), for: .normal)
-        button.titleLabel?.font = UIFont.boldSystemFont(ofSize: 14)
-        button.contentEdgeInsets =  UIEdgeInsets(top: 12, left: 14, bottom: 12, right: 14)
+        button.setTitle(type?.placeholder(), for: .normal)
+        button.titleLabel?.font = UIFont.boldSystemFont(ofSize: 18)
+        button.heightAnchor.constraint(equalToConstant: 49).isActive = true
         return button
     }()
     
@@ -44,41 +36,19 @@ extension UserInfoInputView {
     
     private func configure() {
         self.translatesAutoresizingMaskIntoConstraints = false
-        self.backgroundColor = UIColor.appColor(.gray1)
-        self.layer.cornerRadius = 25
-        self.layer.masksToBounds = true
-        self.heightAnchor.constraint(equalToConstant: 49).isActive = true
     }
     
     private func addSubviews() {
-        self.addSubview(textField)
-        var constraints = [
-            textField.leadingAnchor.constraint(equalTo: self.safeAreaLayoutGuide.leadingAnchor, constant: 20),
-            textField.centerYAnchor.constraint(equalTo: self.safeAreaLayoutGuide.centerYAnchor)
-
-        ]
-        constraints.append(contentsOf: addSubViewButton())
-        NSLayoutConstraint.activate(constraints)
-    }
-    
-    private func addSubViewButton() -> [NSLayoutConstraint] {
-        if let button = self.button {
-            self.addSubview(button)
-            return [
-                button.leadingAnchor.constraint(equalTo: textField.safeAreaLayoutGuide.trailingAnchor,
-                                                constant: 10),
-                button.trailingAnchor.constraint(equalTo: self.safeAreaLayoutGuide.trailingAnchor,
-                                                 constant: -5),
-                button.widthAnchor.constraint(equalToConstant: 100),
-                button.topAnchor.constraint(equalTo: self.safeAreaLayoutGuide.topAnchor,
-                                            constant: 5),
-                button.bottomAnchor.constraint(equalTo: self.safeAreaLayoutGuide.bottomAnchor,
-                                            constant: -5)
-            ]
-        } else {
-            return [textField.trailingAnchor.constraint(equalTo: self.safeAreaLayoutGuide.trailingAnchor,
-                                                    constant: -20)]
-        }
+        self.addSubview(stackView)
+        self.addSubview(signInButton)
+        NSLayoutConstraint.activate([
+            stackView.leadingAnchor.constraint(equalTo: self.safeAreaLayoutGuide.leadingAnchor),
+            stackView.trailingAnchor.constraint(equalTo: self.safeAreaLayoutGuide.trailingAnchor),
+            stackView.topAnchor.constraint(equalTo: self.safeAreaLayoutGuide.topAnchor),
+            signInButton.leadingAnchor.constraint(equalTo: self.safeAreaLayoutGuide.leadingAnchor),
+            signInButton.trailingAnchor.constraint(equalTo: self.safeAreaLayoutGuide.trailingAnchor),
+            signInButton.topAnchor.constraint(equalTo: stackView.safeAreaLayoutGuide.bottomAnchor, constant: 48)
+        ])
     }
     
 }
@@ -86,34 +56,14 @@ extension UserInfoInputView {
 extension UserInfoInputView {
     
     enum InputType {
-        case emailForSignUp, code
-        case email, password, checkPassword, userName
+        case signIn, signUp
         
         func placeholder() -> String {
             switch self {
-            case .emailForSignUp:
-                return "Email - has to be unique"
-            case .code:
-                return "Code"
-            case .email:
-                return "Email"
-            case .password:
-                return "Password"
-            case .checkPassword:
-                return "Check Password "
-            case .userName:
-                return "Username"
-            }
-        }
-        
-        func buttonTitle() -> String {
-            switch self {
-            case .emailForSignUp:
-                return "Send Code"
-            case .code:
-                return "Check Code"
-            default:
-                return ""
+            case .signUp:
+                return "Sign Up"
+            case .signIn:
+                return "Sign In"
             }
         }
     }
